Collection indexes are now built and applied to matching sorts
Query sorts use a 'filter-down' indexing system, finding the best index to use based on the sortData used.

Currently, if an index is defined on a regular & non-unique property, a more specific sort can take a long time to occur (as it uses an insertion sort). This is unavoidable unless using unique indexes (soon).